Add files via upload
Installer batch script file to initially install and get updates from the original main branch on subsequent runs.  Works for Windows and keeps Git for windows up to date in the process.  This is good as initial installer is automated as well to use as an everyday launcher from any location.  This does install to C:\Users\yourusername\breadboard but can be customized as needed by editing the initial start location in the batch file.  This file can be pulled to the main branch if/when desired.   I wrote this to make using the browser easy on a daily basis without having to open a cmd and run the npm manually.